+++ Task038.cs
@@ -1,13 +1,13 @@
-using System.Collections.Generic;
+using System.Numerics;
 using System;
 namespace euler_from26
 {
     public static class Task38
     {
 
-        public static long[] concat_product(long n, long[] p)
+        public static int[] concat_product(int n, int[] p)
         {
-            long[]products = new long[p.Length];
+            int[]products = new int[p.Length];
             for (int i=0; i<p.Length; i++)
                 products[i] = n * p[i];
             return products;
@@ -18,11 +18,11 @@ namespace euler_from26
             for (int i=1; i<10000; i++)
                 for (int n = 2; n <= 9; n++)
                 {
-                    long[] nn = new long[n];
+                    int[] nn = new int[n];
                     for(int j=0; j<n; j++)
                         nn[j] = j+1;
                     var product = concat_product(i, nn);
-                    if (Digits.isPandigital(9, product))
+                    if (Digits.isPandigital((BigInteger)9, product))
                         Console.WriteLine(MyCollections.Print(product));
                 }
             Console.WriteLine();
